Refactor zip generation
